refactor: move lax_numpy indexing routines to their own submodule
